# backend/src/openapi.yaml
openapi: 3.0.3
info:
  title: TaskFlow API Documentation
  version: 1.0.0
  description: >
    REST API for the TaskFlow student project. All successful API responses use
    `{ statusCode, success, message, data }`; errors use
    `{ statusCode, success, message, errors }`.
servers:
  - url: http://localhost:3000/api/v1
    description: Local Development Server
...
paths:
...
  /tasks/{projectId}:
    get:
      tags:
        - Tasks
      summary: List tasks for a project
      description: >
        **Purpose**: Retrieves all tasks belonging to a specified project workspace with populated assignee and assigner details.<br>
        **Authorization**: Required (`Bearer JWT`). User must be a member (`admin`, `project_admin`, `member`).<br>
        **Side Effects**: None.<br>
        **Expected Behavior**: Returns HTTP 200 with all task objects in the project.
      operationId: listTasks
      security:
        - bearerAuth: []
      parameters:
        - name: projectId
          in: path
          required: true
          description: 24-character hex MongoDB ObjectId of the project
          schema:
            type: string
            pattern: '^[0-9a-fA-F]{24}$'
          example: 650c1f1e2f8b9a001c2d3e4f
      responses:
        "200":
          description: Project tasks fetched successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
              example:
                statusCode: 200
                success: true
                message: All tasks related to project fetched successfully!!!
                data:
                  - _id: 650c1f1e2f8b9a001c2d3e88
                    title: Perform Security Audit of JWT Authentication
                    description: Perform comprehensive security audit of JWT token issuance, verification middleware, and cookie storage settings.
                    project: 650c1f1e2f8b9a001c2d3e4f
                    assignedTo:
                      _id: 650c1f1e2f8b9a001c2d3e4f
                      username: alex_rivera
                      email: alex.rivera@taskflow.dev
                    assignedBy:
                      _id: 650c1f1e2f8b9a001c2d3e4f
                      username: alex_rivera
                    status: in_progress
                    attachments:
                      - url: http://localhost:3000/images/security_audit_spec.pdf
                        mimeType: application/pdf
                        size: 1048576
                    createdAt: "2026-01-16T12:00:00.000Z"
        "401":
          $ref: '#/components/responses/Unauthorized'
        "403":
          $ref: '#/components/responses/Forbidden'
        "404":
          $ref: '#/components/responses/NotFound'
        "500":
          $ref: '#/components/responses/InternalServerError'

    post:
      tags:
        - Tasks
      summary: Create a task with optional file attachments
      description: >
        **Purpose**: Creates a new task in a project assigned to a project member, with optional file upload attachments (up to 5 files).<br>
        **Authorization**: Required (`Bearer JWT`). Restricted to `admin` and `project_admin` roles.<br>
        **Side Effects**: Saves file attachments to storage and creates Task record in MongoDB.<br>
        **Expected Behavior**: Returns HTTP 201 with created Task object. Throws HTTP 400 if assigned user is not a project member.
      operationId: createTask
      security:
        - bearerAuth: []
      parameters:
        - name: projectId
          in: path
          required: true
          description: 24-character hex MongoDB ObjectId of the project
          schema:
            type: string
            pattern: '^[0-9a-fA-F]{24}$'
          example: 650c1f1e2f8b9a001c2d3e4f
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/CreateTaskMultipartRequest'
      responses:
        "201":
          description: Task created with uploaded attachments.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
              example:
                statusCode: 201
                success: true
                message: Task created successfully!!!
                data:
                  _id: 650c1f1e2f8b9a001c2d3e88
                  title: Perform Security Audit of JWT Authentication
                  description: Perform comprehensive security audit of JWT token issuance, verification middleware, and cookie storage settings.
                  project: 650c1f1e2f8b9a001c2d3e4f
                  assignedTo: 650c1f1e2f8b9a001c2d3e4f
                  assignedBy: 650c1f1e2f8b9a001c2d3e4f
                  status: todo
                  attachments:
                    - url: http://localhost:3000/images/security_audit_spec.pdf
                      mimeType: application/pdf
                      size: 1048576
                  createdAt: "2026-01-16T12:00:00.000Z"
                  updatedAt: "2026-01-16T12:00:00.000Z"
        "400":
          $ref: '#/components/responses/ValidationError'
        "401":
          $ref: '#/components/responses/Unauthorized'
        "403":
          $ref: '#/components/responses/Forbidden'
        "404":
          $ref: '#/components/responses/NotFound'
        "500":
          $ref: '#/components/responses/InternalServerError'
  /tasks/{projectId}/tasks/{taskId}:
    get:
      tags:
        - Tasks
      summary: Get task details
      description: >
        **Purpose**: Fetches complete task details including subtasks list and populated assignee profile.<br>
        **Authorization**: Required (`Bearer JWT`). User must be a member (`admin`, `project_admin`, `member`).<br>
        **Side Effects**: None.<br>
        **Expected Behavior**: Returns HTTP 200 with aggregated Task object.
      operationId: getTaskById
      security:
        - bearerAuth: []
      parameters:
        - name: projectId
          in: path
          required: true
          description: 24-character hex MongoDB ObjectId of the project
          schema:
            type: string
            pattern: '^[0-9a-fA-F]{24}$'
          example: 650c1f1e2f8b9a001c2d3e4f
        - name: taskId
          in: path
          required: true
          description: 24-character hex MongoDB ObjectId of the task
          schema:
            type: string
            pattern: '^[0-9a-fA-F]{24}$'
          example: 650c1f1e2f8b9a001c2d3e88
      responses:
        "200":
          description: Task details retrieved.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
              example:
                statusCode: 200
                success: true
                message: Task fetched successfully!!!
                data:
                  _id: 650c1f1e2f8b9a001c2d3e88
                  title: Perform Security Audit of JWT Authentication
                  description: Perform comprehensive security audit of JWT token issuance, verification middleware, and cookie storage settings.
                  project: 650c1f1e2f8b9a001c2d3e4f
                  assignedTo:
                    _id: 650c1f1e2f8b9a001c2d3e4f
                    username: alex_rivera
                    fullName: Alex Rivera
                    avatar:
                      url: https://placehold.co/200?text=No+Image
                  status: in_progress
                  subtasks:
                    - _id: 650c1f1e2f8b9a001c2d3e90
                      title: Check token expiration rules
                      isCompleted: true
                  attachments: []
                  createdAt: "2026-01-16T12:00:00.000Z"
        "401":
          $ref: '#/components/responses/Unauthorized'
        "403":
          $ref: '#/components/responses/Forbidden'
        "404":
          $ref: '#/components/responses/NotFound'
        "500":
          $ref: '#/components/responses/InternalServerError'

    put:
      tags:
        - Tasks
      summary: Update task details
      description: >
        **Purpose**: Updates task title, description, status, or assigned user.<br>
        **Authorization**: Required (`Bearer JWT`). Restricted to `admin` and `project_admin` roles.<br>
        **Side Effects**: Updates Task document in MongoDB.<br>
        **Expected Behavior**: Returns HTTP 200 with updated Task object.
      operationId: updateTask
      security:
        - bearerAuth: []
      parameters:
        - name: projectId
          in: path
          required: true
          description: 24-character hex MongoDB ObjectId of the project
          schema:
            type: string
            pattern: '^[0-9a-fA-F]{24}$'
          example: 650c1f1e2f8b9a001c2d3e4f
        - name: taskId
          in: path
          required: true
          description: 24-character hex MongoDB ObjectId of the task to update
          schema:
            type: string
            pattern: '^[0-9a-fA-F]{24}$'
          example: 650c1f1e2f8b9a001c2d3e88
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateTaskRequest'
      responses:
        "200":
          description: Task updated successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
              example:
                statusCode: 200
                success: true
                message: Task details updated successfully!!!
                data:
                  _id: 650c1f1e2f8b9a001c2d3e88
                  title: Perform Security Audit of JWT Authentication
                  description: Perform comprehensive security audit of JWT token issuance, verification middleware, and cookie storage settings.
                  status: done
                  assignedTo: 650c1f1e2f8b9a001c2d3e4f
                  updatedAt: "2026-01-16T15:30:00.000Z"
        "400":
          $ref: '#/components/responses/ValidationError'
        "401":
          $ref: '#/components/responses/Unauthorized'
        "403":
          $ref: '#/components/responses/Forbidden'
        "404":
          $ref: '#/components/responses/NotFound'
        "500":
          $ref: '#/components/responses/InternalServerError'

    delete:
      tags:
        - Tasks
      summary: Delete a task
      description: >
        **Purpose**: Permanently deletes a task and all associated subtasks.<br>
        **Authorization**: Required (`Bearer JWT`). Restricted to `admin` and `project_admin` roles.<br>
        **Side Effects**: Cascade deletes Task and SubTask records from MongoDB.<br>
        **Expected Behavior**: Returns HTTP 200 confirming task deletion.
      operationId: deleteTask
      security:
        - bearerAuth: []
      parameters:
        - name: projectId
          in: path
          required: true
          description: 24-character hex MongoDB ObjectId of the project
          schema:
            type: string
            pattern: '^[0-9a-fA-F]{24}$'
          example: 650c1f1e2f8b9a001c2d3e4f
        - name: taskId
          in: path
          required: true
          description: 24-character hex MongoDB ObjectId of the task to delete
          schema:
            type: string
            pattern: '^[0-9a-fA-F]{24}$'
          example: 650c1f1e2f8b9a001c2d3e88
      responses:
        "200":
          description: Task deleted.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
              example:
                statusCode: 200
                success: true
                message: Task deleted successfully!!!
                data: null
        "401":
          $ref: '#/components/responses/Unauthorized'
        "403":
          $ref: '#/components/responses/Forbidden'
        "404":
          $ref: '#/components/responses/NotFound'
        "500":
          $ref: '#/components/responses/InternalServerError'
  /tasks/{projectId}/tasks/{taskId}/subtasks:
    post:
      tags:
        - Tasks
      summary: Create a subtask
      description: >
        **Purpose**: Creates a new subtask under a parent task.<br>
        **Authorization**: Required (`Bearer JWT`). Restricted to `admin` and `project_admin` roles.<br>
        **Side Effects**: Creates SubTask record in MongoDB.<br>
        **Expected Behavior**: Returns HTTP 201 with created SubTask object.
      operationId: createSubTask
      security:
        - bearerAuth: []
      parameters:
        - name: projectId
          in: path
          required: true
          description: 24-character hex MongoDB ObjectId of the project
          schema:
            type: string
            pattern: '^[0-9a-fA-F]{24}$'
          example: 650c1f1e2f8b9a001c2d3e4f
        - name: taskId
          in: path
          required: true
          description: 24-character hex MongoDB ObjectId of the parent task
          schema:
            type: string
            pattern: '^[0-9a-fA-F]{24}$'
          example: 650c1f1e2f8b9a001c2d3e88
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateSubTaskRequest'
      responses:
        "201":
          description: Subtask created successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
              example:
                statusCode: 201
                success: true
                message: Subtask created successfully!!!
                data:
                  _id: 650c1f1e2f8b9a001c2d3e90
                  title: Check token expiration rules
                  task: 650c1f1e2f8b9a001c2d3e88
                  project: 650c1f1e2f8b9a001c2d3e4f
                  assignedBy: 650c1f1e2f8b9a001c2d3e4f
                  isCompleted: false
                  createdAt: "2026-01-16T13:00:00.000Z"
        "400":
          $ref: '#/components/responses/ValidationError'
        "401":
          $ref: '#/components/responses/Unauthorized'
        "403":
          $ref: '#/components/responses/Forbidden'
        "404":
          $ref: '#/components/responses/NotFound'
        "500":
          $ref: '#/components/responses/InternalServerError'
  /tasks/{projectId}/subtasks/{subTaskId}:
    put:
      tags:
        - Tasks
      summary: Update subtask status or title
      description: >
        **Purpose**: Updates title or completion status (`isCompleted`) of a subtask.<br>
        **Authorization**: Required (`Bearer JWT`). Available to all project members (`admin`, `project_admin`, `member`).<br>
        **Side Effects**: Updates SubTask document in MongoDB.<br>
        **Expected Behavior**: Returns HTTP 200 with updated SubTask object.
      operationId: updateSubTask
      security:
        - bearerAuth: []
      parameters:
        - name: projectId
          in: path
          required: true
          description: 24-character hex MongoDB ObjectId of the project
          schema:
            type: string
            pattern: '^[0-9a-fA-F]{24}$'
          example: 650c1f1e2f8b9a001c2d3e4f
        - name: subTaskId
          in: path
          required: true
          description: 24-character hex MongoDB ObjectId of the subtask
          schema:
            type: string
            pattern: '^[0-9a-fA-F]{24}$'
          example: 650c1f1e2f8b9a001c2d3e90
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateSubTaskRequest'
      responses:
        "200":
          description: Subtask updated successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
              example:
                statusCode: 200
                success: true
                message: SubTask details updated successfully!!!
                data:
                  _id: 650c1f1e2f8b9a001c2d3e90
                  title: Check token expiration rules
                  isCompleted: true
                  task: 650c1f1e2f8b9a001c2d3e88
                  updatedAt: "2026-01-16T14:15:00.000Z"
        "400":
          $ref: '#/components/responses/ValidationError'
        "401":
          $ref: '#/components/responses/Unauthorized'
        "403":
          $ref: '#/components/responses/Forbidden'
        "404":
          $ref: '#/components/responses/NotFound'
        "500":
          $ref: '#/components/responses/InternalServerError'

    delete:
      tags:
        - Tasks
      summary: Delete a subtask
      description: >
        **Purpose**: Removes a subtask from a task.<br>
        **Authorization**: Required (`Bearer JWT`). Restricted to `admin` and `project_admin` roles.<br>
        **Side Effects**: Deletes SubTask document from MongoDB.<br>
        **Expected Behavior**: Returns HTTP 200 confirming subtask deletion.
      operationId: deleteSubTask
      security:
        - bearerAuth: []
      parameters:
        - name: projectId
          in: path
          required: true
          description: 24-character hex MongoDB ObjectId of the project
          schema:
            type: string
            pattern: '^[0-9a-fA-F]{24}$'
          example: 650c1f1e2f8b9a001c2d3e4f
        - name: subTaskId
          in: path
          required: true
          description: 24-character hex MongoDB ObjectId of the subtask
          schema:
            type: string
            pattern: '^[0-9a-fA-F]{24}$'
          example: 650c1f1e2f8b9a001c2d3e90
      responses:
        "200":
          description: Subtask deleted.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
              example:
                statusCode: 200
                success: true
                message: SubTask deleted successfully!!!
                data: null
        "401":
          $ref: '#/components/responses/Unauthorized'
        "403":
          $ref: '#/components/responses/Forbidden'
        "404":
          $ref: '#/components/responses/NotFound'
        "500":
          $ref: '#/components/responses/InternalServerError'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Enter valid JWT Access Token obtained from `/api/v1/auth/login`.
  schemas:
...
    CreateTaskMultipartRequest:
      type: object
      required:
        - title
        - assignedTo
      properties:
        title:
          type: string
          minLength: 5
          example: Perform Security Audit of JWT Authentication
          description: Task title (minimum 5 characters)
        description:
          type: string
          minLength: 20
          nullable: true
          example: Perform comprehensive security audit of JWT token issuance, verification middleware, and cookie storage settings.
          description: Detailed task description (minimum 20 characters)
        assignedTo:
          type: string
          pattern: '^[0-9a-fA-F]{24}$'
          example: 650c1f1e2f8b9a001c2d3e4f
          description: MongoDB ObjectId of the project member assigned to task
        attachments:
          type: array
          description: Optional binary attachment files (up to 5 files)
          items:
            type: string
            format: binary
    UpdateTaskRequest:
      type: object
      properties:
        title:
          type: string
          minLength: 5
          example: Perform Security Audit of JWT Authentication
          description: Updated title
        description:
          type: string
          minLength: 20
          example: Perform comprehensive security audit of JWT token issuance, verification middleware, and cookie storage settings.
          description: Updated description
        status:
          type: string
          enum: [todo, in_progress, done]
          example: done
          description: Updated task status
        assignedTo:
          type: string
          pattern: '^[0-9a-fA-F]{24}$'
          example: 650c1f1e2f8b9a001c2d3e4f
          description: Reassigned user ID
    CreateSubTaskRequest:
      type: object
      required:
        - title
      properties:
        title:
          type: string
          minLength: 5
          example: Check token expiration rules
          description: Subtask title (minimum 5 characters)
    UpdateSubTaskRequest:
      type: object
      properties:
        title:
          type: string
          minLength: 5
          example: Check token expiration rules
          description: Updated subtask title
        isCompleted:
          type: boolean
          example: true
          description: Subtask completion status
...
    ErrorResponse:
      type: object
      required:
        - statusCode
        - success
        - message
        - errors
      properties:
        statusCode:
          type: integer
          example: 400
          description: HTTP response status code
        success:
          type: boolean
          example: false
          description: Success flag (false for errors)
        message:
          type: string
          example: Request failed due to invalid input validation.
          description: Error message summary
        errors:
          type: array
          description: Array of detailed validation error objects or strings
          items:
            type: object
          example:
            - field: email
              message: Invalid email format
        stack:
          type: string
          nullable: true
          example: "Error: Bad Request at..."
          description: Stack trace snippet (only populated in development environment)
    SuccessResponse:
      type: object
      required:
        - statusCode
        - success
        - message
        - data
      properties:
        statusCode:
          type: integer
          example: 200
          description: HTTP response status code
        success:
          type: boolean
          example: true
          description: Success flag (true for successful responses)
        message:
          type: string
          example: Operation completed successfully
          description: Success message summary
        data:
          nullable: true
          description: Response data payload. Its exact type is documented for every endpoint example and in FRONTEND_API_CONTRACT.md.
...
  responses:
    Unauthorized:
      description: Unauthorized - JWT Access Token is missing, invalid, or expired.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            statusCode: 401
            success: false
            message: Unauthorized request or invalid JWT token
            errors: []
    Forbidden:
      description: Forbidden - Insufficient role permissions to access this endpoint or project workspace.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            statusCode: 403
            success: false
            message: You do not have permission to perform this action
            errors: []
    NotFound:
      description: Not Found - The requested resource (user, project, task, subtask, note) does not exist.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            statusCode: 404
            success: false
            message: Resource not found
            errors: []
...
    InternalServerError:
      description: Internal Server Error - An unexpected server error occurred.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            statusCode: 500
            success: false
            message: Internal Server Error
            errors: []
    ValidationError:
      description: Validation Error - Input parameters or payload failed express-validator rules.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            statusCode: 400
            success: false
            message: Validation Error
            errors:
              - field: email
                message: Invalid email format